Remove and simplify a lot of the game/brain argument handling code, in preparation for a move to argparse
