+++ brains/rerun.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python2
 
-import os
 import wave
 from array import array
 
@@ -9,30 +8,19 @@ import pygame
 
 from skeleton_solver import Brain
 
-defaultargs = {'fps': 60,  # run at 60fps because we have a human watching
-               'file': 'output/last_run.pickle',
-               'granularity': 1,  # mostly for converting SuperOpti runs to 60fps
-               'force': False,
-               'recordvideo': False,
-               'recordaudio': False}
-
-validargs = {'fps': lambda x: x >= 0,
-             'file': os.path.isfile,
-             'granularity': lambda x: x > 0}
+default_args = {'file': 'output/last_run.pickle',
+                'force': False,
+                'recordvideo': False,
+                'recordaudio': False}
 
 
 class Rerun(Brain):
     name = 'rerun'
 
-    def __init__(self, game, args=None):
-        Brain.__init__(self, game, args, defaultargs)
+    def __init__(self, game, **kwargs):
+        Brain.__init__(self, game, default_args, **kwargs)
 
-        if args is None:
-            args = {}
-        self.clock = pygame.time.Clock()
-        self.fps = self.args['fps']
         self.force = self.args['force']
-        self.granularity = self.args['granularity']
         self.recordvideo = self.args['recordvideo']
         self.recordaudio = self.args['recordaudio']
 
@@ -52,25 +40,15 @@ class Rerun(Brain):
             if loadedfile['game'] != game.__class__.name:
                 raise Exception('loaded input string is for "%s"' % (loadedfile['game']))
 
-            special_cases = ['granularity', 'audio']
             mismatches = []
             for key in game.args:
-                if key not in special_cases:
+                if key == 'audio':
+                    print('rerun: be sure to use "array" for the game audio if you want to use rerun\'s sound recording.')
+                else:
                     # note: old args being dropped are implicitly ignored by this loop.
                     # explicitly ignore new features with this conditional.
                     if key in loadedfile['game_args'] and loadedfile['game_args'][key] != game.args[key]:
                         mismatches.append(key)
-                else:
-                    if key == 'granularity':
-                        then = loadedfile['game_args'][key]
-                        if key in loadedfile['brain_args']:
-                            then *= loadedfile['brain_args'][key]
-                        now = game.args[key] * self.args[key]
-                        if then != now:
-                            print('rerun: granularity mismatch! consider adjusting rerun\'s granularity.')
-                            mismatches.append(key)
-                    elif key == 'audio':
-                        print('rerun: be sure to use "array" for the game audio if you want to use rerun\'s sound recording.')
             if len(mismatches) > 0:
                 for key in mismatches:
                     print(key, '\n\tgame:', game.args[key], end=' ')
@@ -82,16 +60,11 @@ class Rerun(Brain):
         print('with', len(self.inputstring), 'frames of input')
 
     def Step(self):
-        if self.fps > 0:
-            self.clock.tick(self.fps)
-        frameinput = 0
-
-        if len(self.inputstring):
+        if self.inputstring:
             frameinput = self.inputstring.pop(0)
-
-        for i in range(self.granularity):
-            self.game.Input(frameinput)
             self.outputstring.append(frameinput)
+
+            self.game.Input(frameinput)
             surf = self.game.Draw()
 
             if self.recordvideo:
@@ -101,16 +74,18 @@ class Rerun(Brain):
                                                              len(self.outputstring)))
             if self.recordaudio:
                 self.wav.writeframesraw(array('H', self.game.Sound()).tostring())
-                if self.Victory() and i == self.granularity - 1:
-                    self.wav.close()
 
-            yield surf
+        else:
+            surf = self.game.Draw()
+            self.wav.close()
+
+        yield surf
 
     def Path(self):
-        return self.outputstring
+        return self.outputstring + self.inputstring
 
     def Victory(self):
-        return len(self.inputstring) <= 0
+        return not self.inputstring
 
 
 LoadedBrain = Rerun
