+++ brains/skeleton_solver.py
@@ -1,50 +1,28 @@
 #!/usr/bin/env python2
-
-import pygame
-
-
 class Brain:
     name = 'unnamed solver'
 
-    def __init__(self, game, args=None, defaultargs=None):
-        if defaultargs is None:
-            defaultargs = {}
-        if args is None:
-            args = {}
+    def __init__(self, game, default_args=None, **kwargs):
         self.game = game
-        # try to convert args to appropriate types (from str)
-        for i in args:
-            try:
-                args[i] = eval(args[i], {"__builtins__": None}, {})
-            except:
-                pass
-        # load default values for any keys not given
-        for i in defaultargs:
-            if i not in args:
-                args[i] = defaultargs[i]
-        self.args = args
-        self.terminated = False
+        self.args = (default_args or {}).copy()
+        self.args.update(kwargs)
 
-    # Note: this should 'yield' pygame surfaces throughout execution,
-    #       but it's acceptable to just 'return' a 1-tuple when finished.
-    #       if no screen change was made, return None to skip updating the display
+    # this should 'yield' pygame surfaces throughout execution, or return an iterable of them
     def Step(self):
         for i in self.game.ValidInputs():
             self.game.Input(i)
-        return self.game.Draw(),
+            yield self.game.Draw()
 
     # return the list of input states from start to goal
     def Path(self):
         return []
 
-    # handle events from pygame, if relevant
+    # process pygame events
     def Event(self, evt):
-        if evt.type == pygame.QUIT:
-            self.terminated = True
+        pass
 
     # return the screen (width, height) that should be used
     def ScreenSize(self):
         return self.game.ScreenSize()
-
 
 LoadedBrain = Brain
