+++ games/skeleton_game.py
@@ -10,51 +10,13 @@ ymax = 200  # fallback height of the screen
 class Game:
     name = 'unnamed game'
 
-    def __init__(self, args=None, defaultargs=None, validargs=None):
-        # try to convert args to appropriate types (from str)
-        if args is None:
-            args = {}
-        if defaultargs is None:
-            defaultargs = {}
-        if validargs is None:
-            validargs = {}
-        for i in args:
-            try:
-                args[i] = eval(args[i], {"__builtins__": None}, {})
-            except:
-                pass
-
-        # load default values for any keys not given
-        for i in defaultargs:
-            if i not in args:
-                args[i] = defaultargs[i]
-
-        # check validargs
-        invalids = []
-        for i in args:
-            if i in validargs:
-                validator = validargs[i]
-                try:
-                    iterator = iter(validator)
-                except TypeError:
-                    if hasattr(validator, '__call__'):
-                        if not validator(args[i]):
-                            invalids.append(i)
-                else:
-                    if args[i] not in validator:
-                        print(validator)
-                        invalids.append(i)
-        if len(invalids):
-            for i in invalids:
-                print('invalid value for {}: {}'.format(i, args[i]))
-            raise Exception('bad arguments provided to game.')
-
-        # otherwise, we're all good...
-        self.args = args
+    def __init__(self, default_args=None, **kwargs):
+        self.args = (default_args or {}).copy()
+        self.args.update(kwargs)
 
     # return a copy of the "screen" for visualization
     def Draw(self):
-        ret = pygame.surface.Surface((xmax, ymax))
+        ret = pygame.Surface((xmax, ymax))
         ret.fill((0, 123, 45))
         return ret
 
@@ -80,8 +42,6 @@ class Game:
 
     # return the screen (width, height) that should be used
     def ScreenSize(self):
-        if 'screen' in self.args:
-            return self.args['screen']
         return xmax, ymax
 
     # return left and right channel, 16-bit sound
